==> SimulationStudy/03_FactorVarsEval.Rmd <==
---
title: "FactorVarsEval"
output: html_document
---

Point to correct package library - required on my computer due to R upgrade

```{r}
.libPaths("/home/david/R/x86_64-pc-linux-gnu-library/4.2")
```

The purpose of this script is to identify the number of active factors in simulated datasets.
A factor is active if the variance of cluster parameters (mean) is greater than zero.
This can be used for factor selection in factorization - to give a fair comparison without tuning.

```{r}
library(rjson)
```

Which data to import

```{r}
BaseDir = getwd() # or change it to somewhere else
SimConfigDir = file.path(BaseDir, "SimConfig1")
SimDataDir = file.path(SimConfigDir,"SimData")
```

Import simulation metadata file

```{r}
SimMeta = fromJSON(file = file.path(SimConfigDir,"sim_meta.json"))
LrnCs = 1:SimMeta$CLrn
TrgCs = 1:SimMeta$CTrg+SimMeta$CLrn
```


Loop through the simulations and export files containing number of active factors

```{r}

ActiveKAll = integer()
ActiveKLrn = integer()
ActiveKTrg = integer()

for (i in 1:SimMeta$Sims-1){
  
  SimResDir = file.path(SimDataDir,paste0("sim",i))
  
  ZMuC = as.matrix(read.table(file.path(SimResDir,"ZMuC.csv"),
               header = FALSE, 
               sep=","))
  
  # check Combined set
  CVarAll = as.vector(apply(ZMuC,2,var))
  ActiveKAll = c(ActiveKAll,sum(CVarAll>0))
  
  # check Learning set
  CVarLrn = as.vector(apply(ZMuC[LrnCs,],2,var))
  ActiveKLrn = c(ActiveKLrn,sum(CVarLrn>0))
  
  # check Target set
  CVarTrg = as.vector(apply(ZMuC[TrgCs,],2,var))
  ActiveKTrg = c(ActiveKTrg,sum(CVarTrg>0))
  
}

write.table(ActiveKAll,file=file.path(SimConfigDir,"ActiveKAll.csv"), 
                sep=",", row.names = FALSE, col.names = FALSE)

write.table(ActiveKLrn,file=file.path(SimConfigDir,"ActiveKLrn.csv"), 
                sep=",", row.names = FALSE, col.names = FALSE)

write.table(ActiveKTrg,file=file.path(SimConfigDir,"ActiveKTrg.csv"), 
                sep=",", row.names = FALSE, col.names = FALSE)

```
